rename boost setting
